Remove docs.rs and move everything to appropriate files
