+++ src/download.rs
@@ -1,23 +1,28 @@
-use std::fs::{File, create_dir_all, remove_dir_all, remove_file};
-use std::io::{Read, Write, BufReader, BufWriter};
+use std::fs::{create_dir_all, remove_dir_all, remove_file, File};
+use std::io::{BufReader, BufWriter, Read, Write};
 
 use attohttpc::get;
 
 use flate2::read::GzDecoder;
 use tar::Archive;
 
-use toiletcli::flags::*;
 use toiletcli::flags;
+use toiletcli::flags::*;
 
-use crate::docs::{deserialize_docs_json, Docs};
+use crate::common::{
+    deserialize_docs_json, get_docset_path, get_program_directory, is_docs_json_exists,
+    is_docset_downloaded, is_docset_in_docs,
+};
+use crate::common::{Docs, ResultS};
 
-use crate::common::ResultS;
-use crate::common::{is_docs_json_exists, is_docset_downloaded, is_docset_in_docs, get_docset_path,
-                    get_program_directory};
+use crate::common::{
+    BOLD, DEFAULT_DOWNLOADS_LINK, DEFAULT_USER_AGENT, GREEN, PROGRAM_NAME, RESET, VERSION, YELLOW,
+};
 
-use crate::common::{DEFAULT_DOWNLOADS_LINK, DEFAULT_USER_AGENT, VERSION, BOLD, GREEN, PROGRAM_NAME, RESET, YELLOW};
-
-fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) -> Result<(), String> {
+fn download_docset_tar_gz_with_progress(
+    docset_name: &String,
+    docs: &Vec<Docs>,
+) -> Result<(), String> {
     let user_agent = format!("{DEFAULT_USER_AGENT}/{VERSION}");
 
     for entry in docs.iter() {
@@ -44,7 +49,8 @@ fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) 
                 .send()
                 .map_err(|err| format!("Could not GET {download_link}: {err}"))?;
 
-            let content_length = response.headers()
+            let content_length = response
+                .headers()
                 .get("content-length")
                 .map_or("0", |header| header.to_str().unwrap_or("0"))
                 .parse::<usize>()
@@ -53,7 +59,7 @@ fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) 
             let mut file_writer = BufWriter::new(file);
             let mut response_reader = BufReader::new(response);
 
-            let mut buffer = [0; 1024 * 4];
+            let mut buffer = [0; 1024 * 8];
             let mut file_size = 0;
 
             while let Ok(size) = response_reader.read(&mut buffer) {
@@ -61,23 +67,25 @@ fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) 
                     break;
                 }
 
-                file_writer.write(&buffer[..size])
+                file_writer
+                    .write(&buffer[..size])
                     .map_err(|err| format!("Could not download file: {err}"))?;
 
                 file_size += size;
 
-                print!("\rDownloading `{docset_name}`: received {file_size} of {content_length} bytes...");
+                print!("\rReceived {file_size} of {content_length} bytes...");
             }
             println!();
 
-            file_writer.flush()
+            file_writer
+                .flush()
                 .map_err(|err| format!("Could not flush buffer to file: {err}"))?;
 
             if file_size != content_length {
                 let message = format!(
                     "File size ({file_size}) is different than required size ({content_length}). \
                      Please re-run this command :("
-                    );
+                );
 
                 remove_dir_all(&specific_docset_path)
                     .map_err(|err| format!("Could not remove bad docset ({specific_docset_path:?}): {err}"))?;
@@ -89,7 +97,6 @@ fn download_docset_tar_gz_with_progress(docset_name: &String, docs: &Vec<Docs>) 
 
     Ok(())
 }
-
 
 fn extract_docset_tar_gz(docset_name: &String) -> Result<(), String> {
     let docset_path = get_docset_path(docset_name)?;
@@ -134,7 +141,8 @@ fn show_download_help() -> ResultS {
 
 pub fn download<Args>(mut args: Args) -> ResultS
 where
-    Args: Iterator<Item = String>, {
+    Args: Iterator<Item = String>,
+{
     let mut flag_help;
     let mut flag_force;
 
@@ -152,7 +160,7 @@ where
 
     let docs = deserialize_docs_json()?;
     let mut args_iter = args.iter();
-    let mut success = 0;
+    let mut successful_downloads = 0;
 
     while let Some(docset) = args_iter.next() {
         if !flag_force && is_docset_downloaded(docset)? {
@@ -170,19 +178,19 @@ where
                 continue;
             }
 
-            print!("Downloading `{docset}`...");
+            println!("Downloading `{docset}`...");
             download_docset_tar_gz_with_progress(docset, &docs)?;
 
-            println!("Extracting `{docset}` to `{}`...", get_docset_path(docset)?.display());
+            println!("Extracting to `{}`...", get_docset_path(docset)?.display());
             extract_docset_tar_gz(docset)?;
 
-            success += 1;
+            successful_downloads += 1;
         }
     }
 
-    if success > 1 {
-        println!("{BOLD}{} items were successfully installed{RESET}.", success);
-    } else if success == 1 {
+    if successful_downloads > 1 {
+        println!("{BOLD}{successful_downloads} items were successfully installed{RESET}.");
+    } else if successful_downloads == 1 {
         println!("{BOLD}Install has successfully finished{RESET}.");
     }
