+++ src/list.rs
@@ -1,10 +1,8 @@
-use toiletcli::flags::*;
 use toiletcli::flags;
-
-use crate::docs::deserialize_docs_json;
+use toiletcli::flags::*;
 
 use crate::common::ResultS;
-use crate::common::{is_docs_json_exists, get_local_docsets};
+use crate::common::{deserialize_docs_json, get_local_docsets, is_docs_json_exists};
 use crate::common::{BOLD, GREEN, PROGRAM_NAME, RESET};
 
 fn show_list_help() -> ResultS {
@@ -46,9 +44,7 @@ where
 
     if flag_local {
         let local_docsets = get_local_docsets()?;
-        let mut local_docsets_iter_peekable = local_docsets
-            .iter()
-            .peekable();
+        let mut local_docsets_iter_peekable = local_docsets.iter().peekable();
 
         while let Some(entry) = local_docsets_iter_peekable.next() {
             print!("{GREEN}{} [downloaded]{RESET}", entry);
@@ -64,9 +60,9 @@ where
 
     let docs = deserialize_docs_json()?;
     let docs_names = docs
-            .iter()
-            .map(|entry| entry.slug.to_string())
-            .collect::<Vec<String>>();
+        .iter()
+        .map(|entry| entry.slug.to_string())
+        .collect::<Vec<String>>();
 
     let mut docs_names_peekable = docs_names.iter().peekable();
