+++ src/search.rs
@@ -1,24 +1,27 @@
-use std::fs::File;
-use std::io::{BufWriter, BufReader, Read};
+use std::fs::{read_dir, File};
+use std::io::{BufRead, BufReader, BufWriter, Read};
+use std::path::PathBuf;
 
-use serde::{Serialize, Deserialize};
+use serde::{Deserialize, Serialize};
 use serde_json::{from_str, to_writer};
 
-use toiletcli::flags::*;
 use toiletcli::flags;
+use toiletcli::flags::*;
 
-use crate::{docs::{deserialize_docs_json, print_page_from_docset, search_docset_in_filenames,
-                  search_docset_thoroughly}, debug_println};
+use crate::debug_println;
 
 use crate::common::ResultS;
-use crate::common::{is_docs_json_exists, is_docset_downloaded, is_docset_in_docs, print_search_results,
-                   get_program_directory};
+use crate::common::{
+    convert_paths_to_items, deserialize_docs_json, get_docset_path, get_program_directory,
+    is_docs_json_exists, is_docset_downloaded, is_docset_in_docs, print_page_from_docset,
+    print_search_results,
+};
 use crate::common::{BOLD, GREEN, PROGRAM_NAME, RESET, YELLOW};
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone)]
 struct SearchFlags {
     case_insensitive: bool,
-    precise: bool
+    precise: bool,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -27,7 +30,7 @@ struct SearchCache {
     docset: String,
     exact_items: Vec<String>,
     vague_items: Vec<String>,
-    flags: SearchFlags
+    flags: SearchFlags,
 }
 
 fn try_use_cache(docset: &String, query: &String, flags: &SearchFlags) -> Option<SearchCache> {
@@ -55,13 +58,13 @@ fn write_search_cache(
     query: &String,
     flags: &SearchFlags,
     exact_items: &Vec<String>,
-    vague_items: &Vec<String>) -> ResultS
-{
+    vague_items: &Vec<String>,
+) -> ResultS {
     let program_dir = get_program_directory()?;
     let cache_path = program_dir.join("search_cache.json");
 
-    let cache_file = File::create(&cache_path)
-        .map_err(|err| format!("Could not open {cache_path:?}: {err}"))?;
+    let cache_file =
+        File::create(&cache_path).map_err(|err| format!("Could not open {cache_path:?}: {err}"))?;
 
     let writer = BufWriter::new(cache_file);
 
@@ -70,13 +73,168 @@ fn write_search_cache(
         query:       query.clone(),
         flags:       flags.clone(),
         exact_items: exact_items.clone(),
-        vague_items: vague_items.clone()
+        vague_items: vague_items.clone(),
     };
 
-    to_writer(writer, &cache)
-        .map_err(|err| format!("Could not write cache: {err}"))?;
+    to_writer(writer, &cache).map_err(|err| format!("Could not write cache: {err}"))?;
 
     Ok(())
+}
+
+pub fn search_docset_in_filenames(
+    docset_name: &String,
+    query: &String,
+    case_insensitive: bool,
+) -> Result<Vec<String>, String> {
+    let docset_path = get_docset_path(docset_name)?;
+
+    let internal_query = if case_insensitive {
+        query.to_lowercase()
+    } else {
+        query.to_owned()
+    };
+
+    fn visit_dir_with_query(
+        path: &PathBuf,
+        query: &String,
+        case_insensitive: bool,
+    ) -> Result<Vec<PathBuf>, String> {
+        let mut internal_paths = vec![];
+
+        let dir =
+            read_dir(&path).map_err(|err| format!("Could not read directory {path:?}: {err}"))?;
+
+        for entry in dir {
+            let entry = entry.map_err(|err| format!("Could not read file: {err}"))?;
+
+            let os_file_name = entry.file_name();
+
+            let file_type = entry
+                .file_type()
+                .map_err(|err| format!("Could not read file type of {os_file_name:?}: {err}"))?;
+
+            if file_type.is_dir() {
+                let mut visited = visit_dir_with_query(&entry.path(), &query, case_insensitive)?;
+                internal_paths.append(&mut visited);
+            }
+
+            let mut file_name = os_file_name.to_string_lossy().to_string();
+
+            if file_name.rfind(".html").is_none() {
+                continue;
+            }
+
+            if case_insensitive {
+                file_name.make_ascii_lowercase();
+            }
+
+            if file_name.find(query).is_some() {
+                internal_paths.push(entry.path());
+            }
+        }
+        Ok(internal_paths)
+    }
+
+    let paths = visit_dir_with_query(&docset_path, &internal_query, case_insensitive)?;
+    let mut items = convert_paths_to_items(paths, docset_name)?;
+
+    items.sort_unstable();
+
+    Ok(items)
+}
+
+type ExactMatches = Vec<String>;
+type VagueMatches = Vec<String>;
+
+pub fn search_docset_thoroughly(
+    docset_name: &String,
+    query: &String,
+    case_insensitive: bool,
+) -> Result<(ExactMatches, VagueMatches), String> {
+    let docset_path = get_docset_path(docset_name)?;
+
+    let internal_query = if case_insensitive {
+        query.to_lowercase()
+    } else {
+        query.to_owned()
+    };
+
+    fn visit_dir_with_query(
+        path: &PathBuf,
+        internal_query: &String,
+        case_insensitive: bool,
+    ) -> Result<(Vec<PathBuf>, Vec<PathBuf>), String> {
+        let mut exact_paths = vec![];
+        let mut vague_paths = vec![];
+
+        let dir =
+            read_dir(&path).map_err(|err| format!("Could not read directory {path:?}: {err}"))?;
+
+        for entry in dir {
+            let entry = entry.map_err(|err| format!("Could not read file: {err}"))?;
+
+            let os_file_name = entry.file_name();
+
+            let file_type = entry
+                .file_type()
+                .map_err(|err| format!("Could not read file type of {os_file_name:?}: {err}"))?;
+
+            if file_type.is_dir() {
+                let (mut exact, mut vague) =
+                    visit_dir_with_query(&entry.path(), &internal_query, case_insensitive)?;
+                exact_paths.append(&mut exact);
+                vague_paths.append(&mut vague);
+            }
+
+            let mut file_name = os_file_name.to_string_lossy().to_string();
+
+            if file_name.rfind(".html").is_none() {
+                continue;
+            }
+
+            if case_insensitive {
+                file_name.make_ascii_lowercase();
+            }
+
+            let file_path = entry.path();
+
+            if file_name.find(internal_query).is_some() {
+                exact_paths.push(file_path);
+            } else {
+                let file = File::open(&file_path)
+                    .map_err(|err| format!("Could not open {file_path:?}: {err}"))?;
+                let mut reader = BufReader::new(file);
+                let mut string_buffer = String::new();
+
+                while let Ok(size) = reader.read_line(&mut string_buffer) {
+                    if size == 0 {
+                        break;
+                    }
+
+                    if string_buffer.find(internal_query).is_some() {
+                        vague_paths.push(entry.path());
+                        break;
+                    }
+
+                    string_buffer.clear();
+                }
+            }
+        }
+        Ok((exact_paths, vague_paths))
+    }
+
+    let (exact_paths, vague_paths) =
+        visit_dir_with_query(&docset_path, &internal_query, case_insensitive)?;
+
+    let mut items = (
+        convert_paths_to_items(exact_paths, docset_name)?,
+        convert_paths_to_items(vague_paths, docset_name)?,
+    );
+
+    items.0.sort_unstable();
+    items.1.sort_unstable();
+
+    Ok(items)
 }
 
 fn show_search_help() -> ResultS {
@@ -140,7 +298,10 @@ where
     let flag_open_is_empty = flag_open.is_empty();
     let open_number = flag_open.parse::<usize>().ok();
 
-    let flags = SearchFlags { precise: flag_precise, case_insensitive: flag_case_insensitive };
+    let flags = SearchFlags {
+        precise: flag_precise,
+        case_insensitive: flag_case_insensitive,
+    };
 
     if flag_open_is_empty {
         // Printing query is needed to let you know if you messed up any flags
@@ -195,9 +356,8 @@ where
 
         return Ok(());
     } else {
-        let results =
-        if let Some(cache) = try_use_cache(&docset, &query, &flags) {
-            debug_println!("Using cache");
+        let results = if let Some(cache) = try_use_cache(&docset, &query, &flags) {
+            debug_println!("Search used cache.");
             cache.exact_items
         } else {
             let exact = search_docset_in_filenames(&docset, &query, flag_case_insensitive)?;
@@ -208,14 +368,16 @@ where
         };
 
         if !flag_open_is_empty {
-            if let Some(n) = open_number {
-                if n <= results.len() && n > 0 {
-                    return print_page_from_docset(&docset, &results[n - 1]);
-                } else {
-                    println!("{YELLOW}WARNING{RESET}: --open {n} is out of bounds.");
+            match open_number {
+                Some(n) if n < 1 || n > results.len() => {
+                    println!("{YELLOW}WARNING{RESET}: `--open {n}` is out of bounds.");
                 }
-            } else {
-                println!("{YELLOW}WARNING{RESET}: `--open` requires a number.");
+                Some(n) => {
+                    return print_page_from_docset(&docset, &results[n - 1]);
+                }
+                _ => {
+                    println!("{YELLOW}WARNING{RESET}: `--open` requires a number.");
+                }
             }
         }
